UPDATE config.ini to enable mean missing prices; CHANGE log level to WARNING; MODIFY balance queue error logging; ENHANCE book.py to better handle asset transfers and deduplicate rows; IMPROVE main.py to conditionally rename log file

+++ src/main.py
@@ -61,9 +61,11 @@ def main() -> None:
     # Save log
     log_file_path = evaluation_file_path.with_suffix(".log")
     log_config.shutdown()
-    os.rename(TMP_LOG_FILEPATH, log_file_path)
-
-    print(f"Detailed export saved at {evaluation_file_path} and {log_file_path}")
+    if os.path.exists(TMP_LOG_FILEPATH):
+        os.rename(TMP_LOG_FILEPATH, log_file_path)
+        print(f"Detailed export saved at {evaluation_file_path} and {log_file_path}")
+    else:
+        print(f"Detailed export saved at {evaluation_file_path}")
     print("If you want to archive the evaluation, run `make archive`.")
